Validación JSR-303, Spring con i18n.
Validación de DNI.

Validators.

Internacionalición con vistas y mensajes.

Terminadas Validaciones de cara a Base de Datos y enganchada la capa de
i18n.

+++ GestionDocenteSpring/src/main/java/com/ipartek/formacion/dbms/mappers/ProfesorMapper.java
@@ -15,7 +15,7 @@ public class ProfesorMapper implements RowMapper<Profesor>{
 		profesor.setCodigo(rs.getInt("codigo"));
 		profesor.setNombre(rs.getString("nombre"));
 		profesor.setApellidos(rs.getString("apellidos"));
-		profesor.setnSS(rs.getInt("nss"));
+		profesor.setnSS(rs.getLong("nss"));
 		profesor.setDni(rs.getString("dni"));
 		profesor.setfNacimiento(rs.getDate("fNacimiento"));
